requirements.txt
streamlit
Pillow
pandas
numpy
scikit-learn
plotly
matplotlib
seaborn